fix finding source files when compile_commands is not in root
we'd change the current working directory with chdir to the path of the
compile_commands.json we found.

This is not correct, since if the compile_commands.json file is under
build or below the path's we've been given wouldn't be correct because
they are relative paths and expecting the cwd to be the repository root.

Clang's api also expects to get the directory of where the
compile_commands file is, not the exact path to it. So refactor to use
the directory isntead